requirements.txt
attrs
protobuf==3.20.0
